Move scallingUtils
