+++ src/components/principal/Household.js
@@ -12,7 +12,7 @@ import {
 } from 'react-native';
 import DatePicker from 'react-native-datepicker';
 import AwesomeAlert from 'react-native-awesome-alerts';
-import { scale } from '../scallingUtils';
+import { scale } from '../../utils/scallingUtils';
 import translate from '../../../locales/i18n';
 import { API_URL } from '../../constUtils';
 import ModalSelector from 'react-native-modal-selector';
