+++ src/components/telainicial/Loading.js
@@ -3,7 +3,7 @@ import { ActivityIndicator, AsyncStorage, Image, StatusBar, StyleSheet, View } f
 import { imagemLogo, imagemLogoBR, logoProEpi, logoUnB } from '../../imgs/imageConst';
 import LinearGradient from 'react-native-linear-gradient';
 import translate from '../../../locales/i18n';
-import { scale } from '../scallingUtils';
+import { scale } from '../../utils/scallingUtils';
 import { API_URL } from '../../constUtils';
 
 class AuthLoadingScreen extends React.Component {
