+++ src/components/telainicial/Registrar.js
@@ -11,7 +11,7 @@ import {
 } from 'react-native';
 import DatePicker from 'react-native-datepicker';
 import AwesomeAlert from 'react-native-awesome-alerts';
-import { scale } from '../scallingUtils';
+import { scale } from '../../utils/scallingUtils';
 import translate from '../../../locales/i18n';
 import { API_URL } from '../../constUtils';
 import { CheckBox } from 'react-native-elements';
